pyproject.toml
[project]
name = 'poouli'
version = "0.1.3"
description = 'Syncronize local files with MediaWiki'
authors = [
	{ name = 'Tommi', email = 'surfing@tommi.space' }
]
readme = 'README.md'
license = 'EUPL-1.2'
license-files = ['LICENSE']
requires-python = '>=3.13'
dependencies = [
	'pypandoc-binary>=1.15',
	'python-frontmatter>=1.1.0',
	'pywikibot>=10.6.0',
	'pyyaml>=6.0.3',
	'setuptools>=80.9.0',
]

keywords = [
	'Experimental Publishing',
	'Markdown',
	'MediaWiki',
	'Notes',
	'Pandoc',
	'Wiki',
	'XPUB',
	'Zettelkasten'
]
classifiers = [
	'Development Status :: 3 - Alpha',
	'Environment :: Console',
	'Natural Language :: English',
	'Operating System :: OS Independent',
	'Programming Language :: Python :: 3',
	'Programming Language :: Python :: 3.13',
	'Topic :: Documentation',
	'Topic :: Internet',
	'Topic :: Text Processing',
	'Topic :: Text Processing :: Markup :: Markdown',
	'Topic :: Utilities'
]

[project.urls]
Repository = 'https://codeberg.org/tommi/poouli'
Issues = 'https://codeberg.org/tommi/poouli/issues'
Changelog = 'https://codeberg.org/tommi/poouli/commits/branch/main'

[project.scripts]
poouli = 'poouli:main'

[build-system]
requires = ['uv_build>=0.9.2,<0.10.0']
build-backend = 'uv_build'
